Merge pull request #5 from AstroWedge/fix-money

Fix money

+++ src/main/java/Promotion.java
@@ -1,4 +1,6 @@
+import java.math.BigDecimal;
+
 public interface Promotion
 {
-    double CalculatePromotionTotal(int qtyBought);
+    BigDecimal CalculatePromotionTotal(int qtyBought);
 }
